fix module reference issue
