+++ src/resolvers/refresh_token_registry.py
@@ -2,7 +2,7 @@ import logging
 from datetime import datetime, timezone
 
 from src.mutations.update_token_registry_db import update_token_registry_db_mutation
-from src.services.sifapi.latest_token_registry import latest_token_registry_sifapi
+from src.services.sifapi import *
 from src.utils.setup_logger import setup_logger_util
 
 formatter = logging.Formatter("%(asctime)s-%(name)s-%(levelname)s-%(message)s")
