--- src/services/sifapi/__init__.py
@@ -1,7 +1,0 @@
-# from services.sifapi.get_timestamp_from_height import get_timestamp_from_height_sifapi
-# from services.sifapi.get_timestamp_from_height_pmtp import get_timestamp_from_height_pmtp_sifapi
-# from services.sifapi.get_latest_block_height import get_latest_block_height_sifapi
-# from services.sifapi.get_price_records_pmtp import get_price_records_pmtp_sifapi
-# from services.sifapi.get_price_records import get_price_records_sifapi
-# from services.sifapi.latest_token_registry import latest_token_registry_sifapi
-# from services.sifapi.get_pools import get_pools_sifapi
